backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
Pillow==10.3.0
python-decouple==3.8
requests==2.31.0
yookassa==3.0.0
django-storages==1.14.2
boto3==1.34.69
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.4
flake8==7.0.0
black==24.3.0
isort==5.13.2
python-dateutil==2.9.0
